Fullstack (#2)
* Add db tables for auth and booking

* Switch to typescript init

* Refactor to use typescript

* Add Slack auth

* Add prisma postinstall build step

* set trustHost to true

* Add get booking days endpoint

* Add create booking request. Add get users request

* Add delete booking endpoint

* Changed client to work with new server

* Add loading indicator

---------

+++ src/lib/components/BookedDate.svelte
@@ -1,17 +1,15 @@
-<script>
+<script lang="ts">
   import { createEventDispatcher } from 'svelte';
-  import { css } from '@emotion/css';
-  import { invalidateAll } from '$app/navigation';
-  import { user, parkingSpots } from "$lib/stores/stores.js"
-  import { BookDay, DeleteBooking } from '$lib/Api'
+  import { user, users, bookingDays } from "$lib/stores/stores"
+  import { BookDay, DeleteBooking, GetBookingDays } from '$lib/Api'
   import styles from '$lib/Styles'
   import DateFormat from "$lib/helpers/DateFormat"
-  import BookingDay from "$lib/helpers/BookingDay"
-  import User from '$lib/helpers/User'
+  import BookingUtils from "$lib/helpers/BookingUtils"
   import Button from '$lib/components/Button.svelte';
   import ExpandableButton from "$lib/components/ExpandableButton.svelte";
+	import type { BookingDay } from '$lib/model/models';
 
-  export let bookingDay;
+  export let bookingDay: BookingDay;
   export let style;
 
   let isProcessing = false;
@@ -19,33 +17,40 @@
   const dispatch = createEventDispatcher();
 
   $: formattedDate = DateFormat.localeString(bookingDay.date);
-  $: spotsLeft = BookingDay.parkingSpotsLeft(bookingDay, $parkingSpots);
-  $: haveBooked = bookingDay.bookings.find(booking => booking.userId == $user.id);
-  $: isDefaultUser = User.isDefaultUser($user, $parkingSpots)
+  $: spotsLeft = BookingUtils.parkingSpotsLeft(bookingDay.bookings, $users);
+  $: haveBooked = bookingDay.bookings.find(booking => booking.user.id == $user.id);
 
   const handleClick = () => {
     dispatch('buttonClick');
   };
 
-  const isCancellation = (booking) => {
-    const defaultParkingSpotUsers = $parkingSpots.map(p => p.defaultUserId).filter(id => id > 0);
-    return defaultParkingSpotUsers.includes(booking.userId);
-  }
-
   const bookDay = async () => {
     if (isProcessing) return; 
     isProcessing = true;
-    await BookDay(fetch, $user.id, bookingDay.date);
-    invalidateAll();
+    var newBooking = await BookDay(fetch, { userId: $user.id, date: bookingDay.date, isCancellationBooking: $user.hasPermanentParkingSpot });
+    if (newBooking) {
+      bookingDay.bookings.push(newBooking);
+      const bookingIndex = $bookingDays.findIndex(bd => bd.date == bookingDay.date);
+      $bookingDays[bookingIndex] = bookingDay;
+    } else {
+      $bookingDays = await GetBookingDays(fetch);
+    }
+    
     isProcessing = false;
   }
 
   const deleteBooking = async () => {
     if (isProcessing) return; 
     isProcessing = true;
-    const bookingToDelete = bookingDay.bookings.find(booking => booking.userId == $user.id);
+    const bookingToDelete = bookingDay.bookings.find(booking => booking.user.id == $user.id);
+    if (!bookingToDelete) {
+      isProcessing = false;
+      return;
+    };
     await DeleteBooking(fetch, bookingToDelete.id);
-    invalidateAll();
+    bookingDay.bookings = bookingDay.bookings.filter(booking => booking.id != bookingToDelete.id);
+    const bookingIndex = $bookingDays.findIndex(bd => bd.date == bookingDay.date);
+    $bookingDays[bookingIndex] = bookingDay;
     isProcessing = false;
   }
 
@@ -58,18 +63,18 @@
     </div>
     <div slot="expanded">
       {#each bookingDay.bookings as booking (booking.id)}
-        <p class="name" style="text-decoration: {isCancellation(booking) ? 'line-through' : 'none'};">{booking.userName}</p>
+        <p class="name" style="text-decoration: {booking.isCancellationBooking ? 'line-through' : 'none'};">{booking.user.name}</p>
       {/each}
       {#if haveBooked}
         <div class="book-button">
-          <Button style={{...styles.button.primary, padding: '0 15px'}} on:buttonClick={deleteBooking}>
-            {isDefaultUser ? 'Fjern kansellering' : 'Fjern reservasjon'}
+          <Button style={{...styles.button.primary, padding: '0 15px', minWidth: '185px'}} on:buttonClick={deleteBooking} loading={isProcessing}>
+            {haveBooked.isCancellationBooking ? 'Fjern kansellering' : 'Fjern reservasjon'}
           </Button>
         </div>
       {/if}
       {#if !haveBooked && spotsLeft > 0}
         <div class="book-button">
-          <Button style={{...styles.button.primary, padding: '0 15px'}} on:buttonClick={bookDay}>
+          <Button style={{...styles.button.primary, padding: '0 15px', minWidth: '185px'}} on:buttonClick={bookDay} loading={isProcessing}>
             Book
           </Button>
         </div>
